+++ package.json
@@ -2,6 +2,8 @@
   "name": "app",
   "private": true,
   "devDependencies": {
+    "buffer": "^6.0.3",
+    "process": "^0.11.10",
     "webpack": "^4.46.0",
     "webpack-cli": "^3.3.12",
     "webpack-dev-server": "^3"
